feat: refactor code to enable multiple environments
