+++ lib/async-hooks/index.ts
@@ -1,3 +1,2 @@
 export * from "./async-context";
-export * from "./async-hooks-middleware";
 export * from "./async-hooks-module";
