+++ lib/clients/http/axios-tracing.interceptor.ts
@@ -2,8 +2,8 @@ import { HttpService, Injectable, OnModuleInit } from "@nestjs/common";
 import { getCauseTypeFromHttpStatus } from "aws-xray-sdk-core/lib/utils";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
 import { ClientRequest, IncomingMessage } from "http";
-import { TracingNotInitializedException } from "../exceptions";
-import { TracingService } from "../tracing.service";
+import { TracingService } from "../../core";
+import { TracingNotInitializedException } from "../../exceptions";
 import { HEADER_TRACE_CONTEXT } from "./http-tracing.constants";
 
 export type AxiosOnFulfilledInterceptor<T> = (value: T) => T | Promise<T>;
