+++ lib/core/index.ts
@@ -1,0 +1,2 @@
+export * from "./core.module";
+export * from "./tracing.service";
