+++ lib/core/interfaces/index.ts
