+++ lib/core/interfaces/tracing-config.interface.ts
@@ -12,7 +12,10 @@ export class TracingConfig {
   public rate?: number;
 
   /**
-   * Address of X-Ray Daemon
+   * Address of X-Ray Daemon.
+   *
+   * This value can only be set through the environment variable `AWS_XRAY_DAEMON_ADDRESS`.
+   * This is only documented in the README in section "Known Bugs".
    */
   public daemonAddress?: string;
 }
