+++ lib/core/interfaces/xray-client.interface.ts
