+++ lib/environments/http/index.ts
@@ -1,0 +1,1 @@
+export { HttpEnvironmentModule } from "./http-environment.module";
